Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,10 +2,10 @@
 
 int main() {
     int x,y,z;
-    int avg;
+    float avg;
     scanf("%d %d %d",&x,&y,&z);
     avg=(x+y+z)/3;
-    printf("Average:%d",avg);
+    printf("Average:%.2f",avg);
 
    
     return 0;
